requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI & LLM
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.5
langgraph==0.0.20

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Cache & Queue
redis==5.0.1
celery==5.3.6

# Scheduler (P0 Autopilot)
APScheduler==3.10.4

# HTTP & API
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# WordPress Integration
python-wordpress-xmlrpc==2.3

# SEO & Keyword Research
beautifulsoup4==4.12.3
lxml==5.1.0

# Google API (P1 - GSC Integration)
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Image Processing
Pillow==10.2.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
pyyaml==6.0.1
jinja2==3.1.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
ruff==0.1.14
mypy==1.8.0
ruff==0.1.14
